+++ src/app/vendors/d2-vendors.test.ts
@@ -1,5 +1,3 @@
-/* TODO: Reenable after vendors comes back online after updating vendors profile
-
 import { getBuckets } from 'app/destiny2/d2-buckets';
 import { getTestDefinitions, getTestProfile, getTestVendors } from 'testing/test-utils';
 import { D2VendorGroup, toVendorGroups } from './d2-vendors';
@@ -49,14 +47,5 @@ describe('process vendors', () => {
       }
     }
     expect(vendorItemPatternFound).toBe(true);
-  });
-});
-*/
-
-// a test is required...
-describe('vendors offline', () => {
-  it('vendors are offline', async () => {
-    let vendors = false;
-    expect(vendors).toBe(false);
   });
 });